+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteToolBar/ModalConfirmDownloadUnsafe.tsx
@@ -19,7 +19,7 @@ export const ModalConfirmDownloadUnsafe = ({
       isOpen
       closeOnClickOutside
       onClose={() => onClose()}
-      aria-describedby="modal-confirm-download-unsafe-title"
+      aria-label={t('Warning')}
       rightActions={
         <>
           <Button
@@ -48,7 +48,7 @@ export const ModalConfirmDownloadUnsafe = ({
       size={ModalSize.SMALL}
       title={
         <Text
-          as="h1"
+          as="h2"
           id="modal-confirm-download-unsafe-title"
           $gap="0.7rem"
           $size="h6"
@@ -61,10 +61,7 @@ export const ModalConfirmDownloadUnsafe = ({
         </Text>
       }
     >
-      <Box
-        aria-label={t('Modal confirmation to download the attachment')}
-        className="--docs--modal-confirm-download-unsafe"
-      >
+      <Box className="--docs--modal-confirm-download-unsafe">
         <Box>
           <Box $direction="column" $gap="0.35rem" $margin={{ top: 'sm' }}>
             <Text $variation="secondary">